+++ cslib/Erlang/Int.cs
@@ -5,6 +5,6 @@ namespace CsLib.Erlang
 {
   public sealed class Int : Term
   {
-    internal Int(Runtime runtime, int handle) : base(runtime, handle) {}
+    internal Int(Runtime runtime, Int64 handle) : base(runtime, handle) {}
   }
 }
