+++ cslib/Erlang/Pid.cs
@@ -6,6 +6,6 @@ namespace CsLib.Erlang
 {
   public sealed class Pid : Term
   {
-    internal Pid(Runtime runtime, int handle) : base(runtime, handle) {}
+    internal Pid(Runtime runtime, Int64 handle) : base(runtime, handle) {}
   }
 }
