+++ cslib/Erlang/PointerResource.cs
@@ -6,6 +6,6 @@ namespace CsLib.Erlang
 {
   public sealed class PointerResource : Term
   {
-    internal PointerResource(Runtime runtime, int handle) : base(runtime, handle) {}
+    internal PointerResource(Runtime runtime, Int64 handle) : base(runtime, handle) {}
   }
 }
