+++ cslib/Erlang/ProcessResult.cs
@@ -5,6 +5,6 @@ namespace CsLib.Erlang
 {
   public sealed class ProcessResult : Term
   {
-    internal ProcessResult(Runtime runtime, int handle) : base(runtime, handle) {}
+    internal ProcessResult(Runtime runtime, Int64 handle) : base(runtime, handle) {}
   }
 }
