+++ cslib/Erlang/Tuple.cs
@@ -5,6 +5,6 @@ namespace CsLib.Erlang
 {
   public sealed class Tuple : Term
   {
-    internal Tuple(Runtime runtime, int handle) : base(runtime, handle) {}
+    internal Tuple(Runtime runtime, Int64 handle) : base(runtime, handle) {}
   }
 }
